+++ subsys/bluetooth/shell/CMakeLists.txt
@@ -7,6 +7,10 @@ zephyr_library_sources_ifdef(
   gatt.c
   )
 zephyr_library_sources_ifdef(
+  CONFIG_BT_BREDR
+  bredr.c
+  )
+zephyr_library_sources_ifdef(
   CONFIG_BT_CTLR
   ll.c
   ticker.c
