+++ subsys/bluetooth/shell/bt.c
@@ -38,7 +38,6 @@
 #define DEVICE_NAME_LEN		(sizeof(DEVICE_NAME) - 1)
 #define CREDITS			10
 #define DATA_MTU		(23 * CREDITS)
-#define DATA_BREDR_MTU		48
 
 static u8_t selected_id = BT_ID_DEFAULT;
 const struct shell *ctx_shell;
@@ -54,83 +53,6 @@ static struct bt_conn *pairing_conn;
 NET_BUF_POOL_DEFINE(data_tx_pool, 1, DATA_MTU, BT_BUF_USER_DATA_MIN, NULL);
 NET_BUF_POOL_DEFINE(data_rx_pool, 1, DATA_MTU, BT_BUF_USER_DATA_MIN, NULL);
 #endif
-
-#if defined(CONFIG_BT_BREDR)
-NET_BUF_POOL_DEFINE(data_bredr_pool, 1, DATA_BREDR_MTU, BT_BUF_USER_DATA_MIN,
-		    NULL);
-
-#define SDP_CLIENT_USER_BUF_LEN		512
-NET_BUF_POOL_DEFINE(sdp_client_pool, CONFIG_BT_MAX_CONN,
-		    SDP_CLIENT_USER_BUF_LEN, BT_BUF_USER_DATA_MIN, NULL);
-#endif /* CONFIG_BT_BREDR */
-
-#if defined(CONFIG_BT_RFCOMM)
-
-static struct bt_sdp_attribute spp_attrs[] = {
-	BT_SDP_NEW_SERVICE,
-	BT_SDP_LIST(
-		BT_SDP_ATTR_SVCLASS_ID_LIST,
-		BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 3),
-		BT_SDP_DATA_ELEM_LIST(
-		{
-			BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-			BT_SDP_ARRAY_16(BT_SDP_SERIAL_PORT_SVCLASS)
-		},
-		)
-	),
-	BT_SDP_LIST(
-		BT_SDP_ATTR_PROTO_DESC_LIST,
-		BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 12),
-		BT_SDP_DATA_ELEM_LIST(
-		{
-			BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 3),
-			BT_SDP_DATA_ELEM_LIST(
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-				BT_SDP_ARRAY_16(BT_SDP_PROTO_L2CAP)
-			},
-			)
-		},
-		{
-			BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 5),
-			BT_SDP_DATA_ELEM_LIST(
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-				BT_SDP_ARRAY_16(BT_SDP_PROTO_RFCOMM)
-			},
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UINT8),
-				BT_SDP_ARRAY_8(BT_RFCOMM_CHAN_SPP)
-			},
-			)
-		},
-		)
-	),
-	BT_SDP_LIST(
-		BT_SDP_ATTR_PROFILE_DESC_LIST,
-		BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 8),
-		BT_SDP_DATA_ELEM_LIST(
-		{
-			BT_SDP_TYPE_SIZE_VAR(BT_SDP_SEQ8, 6),
-			BT_SDP_DATA_ELEM_LIST(
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UUID16),
-				BT_SDP_ARRAY_16(BT_SDP_SERIAL_PORT_SVCLASS)
-			},
-			{
-				BT_SDP_TYPE_SIZE(BT_SDP_UINT16),
-				BT_SDP_ARRAY_16(0x0102)
-			},
-			)
-		},
-		)
-	),
-	BT_SDP_SERVICE_NAME("Serial Port"),
-};
-
-static struct bt_sdp_record spp_rec = BT_SDP_RECORD(spp_attrs);
-
-#endif /* CONFIG_BT_RFCOMM */
 
 #define NAME_LEN 30
 
@@ -197,7 +119,7 @@ static const char *current_prompt(void)
 }
 #endif
 
-static void conn_addr_str(struct bt_conn *conn, char *addr, size_t len)
+void conn_addr_str(struct bt_conn *conn, char *addr, size_t len)
 {
 	struct bt_conn_info info;
 
@@ -272,132 +194,6 @@ static void le_param_updated(struct bt_conn *conn, u16_t interval,
 	print(NULL, "LE conn param updated: int 0x%04x lat %d "
 		     "to %d\n", interval, latency, timeout);
 }
-
-#if defined(CONFIG_BT_BREDR)
-static u8_t sdp_hfp_ag_user(struct bt_conn *conn,
-			       struct bt_sdp_client_result *result)
-{
-	char addr[BT_ADDR_STR_LEN];
-	u16_t param, version;
-	u16_t features;
-	int res;
-
-	conn_addr_str(conn, addr, sizeof(addr));
-
-	if (result) {
-		printk("SDP HFPAG data@%p (len %u) hint %u from remote %s\n",
-			result->resp_buf, result->resp_buf->len,
-			result->next_record_hint, addr);
-
-		/*
-		 * Focus to get BT_SDP_ATTR_PROTO_DESC_LIST attribute item to
-		 * get HFPAG Server Channel Number operating on RFCOMM protocol.
-		 */
-		res = bt_sdp_get_proto_param(result->resp_buf,
-					     BT_SDP_PROTO_RFCOMM, &param);
-		if (res < 0) {
-			printk("Error getting Server CN, err %d\n", res);
-			goto done;
-		}
-		printk("HFPAG Server CN param 0x%04x\n", param);
-
-		res = bt_sdp_get_profile_version(result->resp_buf,
-						 BT_SDP_HANDSFREE_SVCLASS,
-						 &version);
-		if (res < 0) {
-			printk("Error getting profile version, err %d\n", res);
-			goto done;
-		}
-		printk("HFP version param 0x%04x\n", version);
-
-		/*
-		 * Focus to get BT_SDP_ATTR_SUPPORTED_FEATURES attribute item to
-		 * get profile Supported Features mask.
-		 */
-		res = bt_sdp_get_features(result->resp_buf, &features);
-		if (res < 0) {
-			printk("Error getting HFPAG Features, err %d\n", res);
-			goto done;
-		}
-		printk("HFPAG Supported Features param 0x%04x\n", features);
-	} else {
-		printk("No SDP HFPAG data from remote %s\n", addr);
-	}
-done:
-	return BT_SDP_DISCOVER_UUID_CONTINUE;
-}
-
-static u8_t sdp_a2src_user(struct bt_conn *conn,
-			      struct bt_sdp_client_result *result)
-{
-	char addr[BT_ADDR_STR_LEN];
-	u16_t param, version;
-	u16_t features;
-	int res;
-
-	conn_addr_str(conn, addr, sizeof(addr));
-
-	if (result) {
-		printk("SDP A2SRC data@%p (len %u) hint %u from remote %s\n",
-			result->resp_buf, result->resp_buf->len,
-			result->next_record_hint, addr);
-
-		/*
-		 * Focus to get BT_SDP_ATTR_PROTO_DESC_LIST attribute item to
-		 * get A2SRC Server PSM Number.
-		 */
-		res = bt_sdp_get_proto_param(result->resp_buf,
-					     BT_SDP_PROTO_L2CAP, &param);
-		if (res < 0) {
-			printk("A2SRC PSM Number not found, err %d\n", res);
-			goto done;
-		}
-		printk("A2SRC Server PSM Number param 0x%04x\n", param);
-
-		/*
-		 * Focus to get BT_SDP_ATTR_PROFILE_DESC_LIST attribute item to
-		 * get profile version number.
-		 */
-		res = bt_sdp_get_profile_version(result->resp_buf,
-						 BT_SDP_ADVANCED_AUDIO_SVCLASS,
-						 &version);
-		if (res < 0) {
-			printk("A2SRC version not found, err %d\n", res);
-			goto done;
-		}
-		printk("A2SRC version param 0x%04x\n", version);
-
-		/*
-		 * Focus to get BT_SDP_ATTR_SUPPORTED_FEATURES attribute item to
-		 * get profile supported features mask.
-		 */
-		res = bt_sdp_get_features(result->resp_buf, &features);
-		if (res < 0) {
-			printk("A2SRC Features not found, err %d\n", res);
-			goto done;
-		}
-		printk("A2SRC Supported Features param 0x%04x\n", features);
-	} else {
-		printk("No SDP A2SRC data from remote %s\n", addr);
-	}
-done:
-	return BT_SDP_DISCOVER_UUID_CONTINUE;
-}
-
-static struct bt_sdp_discover_params discov_hfpag = {
-	.uuid = BT_UUID_DECLARE_16(BT_SDP_HANDSFREE_AGW_SVCLASS),
-	.func = sdp_hfp_ag_user,
-	.pool = &sdp_client_pool,
-};
-
-static struct bt_sdp_discover_params discov_a2src = {
-	.uuid = BT_UUID_DECLARE_16(BT_SDP_AUDIO_SOURCE_SVCLASS),
-	.func = sdp_a2src_user,
-	.pool = &sdp_client_pool,
-};
-
-static struct bt_sdp_discover_params discov;
-#endif
 
 #if defined(CONFIG_BT_SMP)
 static void identity_resolved(struct bt_conn *conn, const bt_addr_le_t *rpa,
@@ -482,7 +278,7 @@ static int hexstr2array(const char *str, u8_t *array, u8_t size)
 	return 0;
 }
 
-static int str2bt_addr(const char *str, bt_addr_t *addr)
+int str2bt_addr(const char *str, bt_addr_t *addr)
 {
 	return hexstr2array(str, addr->val, 6);
 }
@@ -1498,180 +1294,6 @@ static void cmd_auth_passkey(const struct shell *shell,
 }
 #endif /* CONFIG_BT_SMP) || CONFIG_BT_BREDR */
 
-#if defined(CONFIG_BT_BREDR)
-static void cmd_auth_pincode(const struct shell *shell,
-			    size_t argc, char *argv[])
-{
-	struct bt_conn *conn;
-	u8_t max = 16;
-
-	if (default_conn) {
-		conn = default_conn;
-	} else if (pairing_conn) {
-		conn = pairing_conn;
-	} else {
-		conn = NULL;
-	}
-
-	if (!conn) {
-		printk("Not connected\n");
-		return;
-	}
-
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	if (strlen(argv[1]) > max) {
-		printk("PIN code value invalid - enter max %u digits\n", max);
-		return;
-	}
-
-	printk("PIN code \"%s\" applied\n", argv[1]);
-
-	bt_conn_auth_pincode_entry(conn, argv[1]);
-}
-
-static void cmd_connect_bredr(const struct shell *shell,
-			     size_t argc, char *argv[])
-{
-	struct bt_conn *conn;
-	bt_addr_t addr;
-	int err;
-
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	err = str2bt_addr(argv[1], &addr);
-	if (err) {
-		printk("Invalid peer address (err %d)\n", err);
-		return;
-	}
-
-	conn = bt_conn_create_br(&addr, BT_BR_CONN_PARAM_DEFAULT);
-	if (!conn) {
-		printk("Connection failed\n");
-	} else {
-
-		printk("Connection pending\n");
-
-		/* unref connection obj in advance as app user */
-		bt_conn_unref(conn);
-	}
-
-	return;
-}
-
-static void br_device_found(const bt_addr_t *addr, s8_t rssi,
-				  const u8_t cod[3], const u8_t eir[240])
-{
-	char br_addr[BT_ADDR_STR_LEN];
-	char name[239];
-	int len = 240;
-
-	(void)memset(name, 0, sizeof(name));
-
-	while (len) {
-		if (len < 2) {
-			break;
-		};
-
-		/* Look for early termination */
-		if (!eir[0]) {
-			break;
-		}
-
-		/* Check if field length is correct */
-		if (eir[0] > len - 1) {
-			break;
-		}
-
-		switch (eir[1]) {
-		case BT_DATA_NAME_SHORTENED:
-		case BT_DATA_NAME_COMPLETE:
-			if (eir[0] > sizeof(name) - 1) {
-				memcpy(name, &eir[2], sizeof(name) - 1);
-			} else {
-				memcpy(name, &eir[2], eir[0] - 1);
-			}
-			break;
-		default:
-			break;
-		}
-
-		/* Parse next AD Structure */
-		len -= eir[0] + 1;
-		eir += eir[0] + 1;
-	}
-
-	bt_addr_to_str(addr, br_addr, sizeof(br_addr));
-
-	printk("[DEVICE]: %s, RSSI %i %s\n", br_addr, rssi, name);
-}
-
-static struct bt_br_discovery_result br_discovery_results[5];
-
-static void br_discovery_complete(struct bt_br_discovery_result *results,
-				  size_t count)
-{
-	size_t i;
-
-	printk("BR/EDR discovery complete\n");
-
-	for (i = 0; i < count; i++) {
-		br_device_found(&results[i].addr, results[i].rssi,
-				results[i].cod, results[i].eir);
-	}
-}
-
-static void cmd_bredr_discovery(const struct shell *shell,
-				size_t argc, char *argv[])
-{
-	const char *action;
-
-	if (!shell_cmd_precheck(shell, argc >= 2, NULL, 0)) {
-		return;
-	}
-
-	action = argv[1];
-	if (!strcmp(action, "on")) {
-		struct bt_br_discovery_param param;
-
-		param.limited = false;
-		param.length = 8;
-
-		if (argc > 2) {
-			param.length = atoi(argv[2]);
-		}
-
-		if (argc > 3 && !strcmp(argv[3], "limited")) {
-			param.limited = true;
-		}
-
-		if (bt_br_discovery_start(&param, br_discovery_results,
-					  ARRAY_SIZE(br_discovery_results),
-					  br_discovery_complete) < 0) {
-			printk("Failed to start discovery\n");
-			return;
-		}
-
-		printk("Discovery started\n");
-	} else if (!strcmp(action, "off")) {
-		if (bt_br_discovery_stop()) {
-			printk("Failed to stop discovery\n");
-			return;
-		}
-
-		printk("Discovery stopped\n");
-	} else {
-		shell_help_print(shell, NULL, 0);
-		return;
-	}
-}
-
-#endif /* CONFIG_BT_BREDR */
-
 #if defined(CONFIG_BT_L2CAP_DYNAMIC_CHANNEL)
 static u32_t l2cap_rate;
 static u32_t l2cap_recv_delay;
@@ -1932,335 +1554,6 @@ static void cmd_l2cap_metrics(const struct shell *shell,
 
 #endif
 
-#if defined(CONFIG_BT_BREDR)
-static int l2cap_bredr_recv(struct bt_l2cap_chan *chan, struct net_buf *buf)
-{
-	printk("Incoming data channel %p len %u\n", chan, buf->len);
-
-	return 0;
-}
-
-static void l2cap_bredr_connected(struct bt_l2cap_chan *chan)
-{
-	printk("Channel %p connected\n", chan);
-}
-
-static void l2cap_bredr_disconnected(struct bt_l2cap_chan *chan)
-{
-	printk("Channel %p disconnected\n", chan);
-}
-
-static struct net_buf *l2cap_bredr_alloc_buf(struct bt_l2cap_chan *chan)
-{
-	printk("Channel %p requires buffer\n", chan);
-
-	return net_buf_alloc(&data_bredr_pool, K_FOREVER);
-}
-
-static struct bt_l2cap_chan_ops l2cap_bredr_ops = {
-	.alloc_buf	= l2cap_bredr_alloc_buf,
-	.recv		= l2cap_bredr_recv,
-	.connected	= l2cap_bredr_connected,
-	.disconnected	= l2cap_bredr_disconnected,
-};
-
-static struct bt_l2cap_br_chan l2cap_bredr_chan = {
-	.chan.ops	= &l2cap_bredr_ops,
-	 /* Set for now min. MTU */
-	.rx.mtu		= 48,
-};
-
-static int l2cap_bredr_accept(struct bt_conn *conn, struct bt_l2cap_chan **chan)
-{
-	printk("Incoming BR/EDR conn %p\n", conn);
-
-	if (l2cap_bredr_chan.chan.conn) {
-		printk("No channels available");
-		return -ENOMEM;
-	}
-
-	*chan = &l2cap_bredr_chan.chan;
-
-	return 0;
-}
-
-static struct bt_l2cap_server br_server = {
-	.accept = l2cap_bredr_accept,
-};
-
-static void cmd_bredr_l2cap_register(const struct shell *shell,
-				    size_t argc, char *argv[])
-{
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	if (br_server.psm) {
-		printk("Already registered\n");
-		return;
-	}
-
-	br_server.psm = strtoul(argv[1], NULL, 16);
-
-	if (bt_l2cap_br_server_register(&br_server) < 0) {
-		printk("Unable to register psm\n");
-		br_server.psm = 0;
-	} else {
-		printk("L2CAP psm %u registered\n", br_server.psm);
-	}
-}
-
-#if defined(CONFIG_BT_RFCOMM)
-static void rfcomm_bredr_recv(struct bt_rfcomm_dlc *dlci, struct net_buf *buf)
-{
-	printk("Incoming data dlc %p len %u\n", dlci, buf->len);
-}
-
-static void rfcomm_bredr_connected(struct bt_rfcomm_dlc *dlci)
-{
-	printk("Dlc %p connected\n", dlci);
-}
-
-static void rfcomm_bredr_disconnected(struct bt_rfcomm_dlc *dlci)
-{
-	printk("Dlc %p disconnected\n", dlci);
-}
-
-static struct bt_rfcomm_dlc_ops rfcomm_bredr_ops = {
-	.recv		= rfcomm_bredr_recv,
-	.connected	= rfcomm_bredr_connected,
-	.disconnected	= rfcomm_bredr_disconnected,
-};
-
-static struct bt_rfcomm_dlc rfcomm_dlc = {
-	.ops = &rfcomm_bredr_ops,
-	.mtu = 30,
-};
-
-static int rfcomm_bredr_accept(struct bt_conn *conn, struct bt_rfcomm_dlc **dlc)
-{
-	printk("Incoming RFCOMM conn %p\n", conn);
-
-	if (rfcomm_dlc.session) {
-		printk("No channels available");
-		return -ENOMEM;
-	}
-
-	*dlc = &rfcomm_dlc;
-
-	return 0;
-}
-
-struct bt_rfcomm_server rfcomm_server = {
-	.accept = &rfcomm_bredr_accept,
-};
-
-static void cmd_bredr_rfcomm_register(const struct shell *shell,
-				      size_t argc, char *argv[])
-{
-	int ret;
-
-	if (rfcomm_server.channel) {
-		printk("Already registered\n");
-		return;
-	}
-
-	rfcomm_server.channel = BT_RFCOMM_CHAN_SPP;
-
-	ret = bt_rfcomm_server_register(&rfcomm_server);
-	if (ret < 0) {
-		printk("Unable to register channel %x\n", ret);
-		rfcomm_server.channel = 0;
-	} else {
-		printk("RFCOMM channel %u registered\n", rfcomm_server.channel);
-		bt_sdp_register_service(&spp_rec);
-	}
-}
-
-static void cmd_rfcomm_connect(const struct shell *shell,
-			      size_t argc, char *argv[])
-{
-	u8_t channel;
-	int err;
-
-	if (!default_conn) {
-		printk("Not connected\n");
-		return;
-	}
-
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	channel = strtoul(argv[1], NULL, 16);
-
-	err = bt_rfcomm_dlc_connect(default_conn, &rfcomm_dlc, channel);
-	if (err < 0) {
-		printk("Unable to connect to channel %d (err %u)\n",
-		       channel, err);
-	} else {
-		printk("RFCOMM connection pending\n");
-	}
-}
-
-static void cmd_rfcomm_send(const struct shell *shell,
-			    size_t argc, char *argv[])
-{
-	u8_t buf_data[DATA_BREDR_MTU] = { [0 ... (DATA_BREDR_MTU - 1)] =
-					    0xff };
-	int ret, len, count = 1;
-	struct net_buf *buf;
-
-	if (argc > 1) {
-		count = strtoul(argv[1], NULL, 10);
-	}
-
-	while (count--) {
-		buf = bt_rfcomm_create_pdu(&data_bredr_pool);
-		/* Should reserve one byte in tail for FCS */
-		len = min(rfcomm_dlc.mtu, net_buf_tailroom(buf) - 1);
-
-		net_buf_add_mem(buf, buf_data, len);
-		ret = bt_rfcomm_dlc_send(&rfcomm_dlc, buf);
-		if (ret < 0) {
-			printk("Unable to send: %d\n", -ret);
-			net_buf_unref(buf);
-			break;
-		}
-	}
-}
-
-static void cmd_rfcomm_disconnect(const struct shell *shell,
-				 size_t argc, char *argv[])
-{
-	int err;
-
-	err = bt_rfcomm_dlc_disconnect(&rfcomm_dlc);
-	if (err) {
-		printk("Unable to disconnect: %u\n", -err);
-	}
-}
-
-#endif /* CONFIG_BT_RFCOMM) */
-
-static void cmd_bredr_discoverable(const struct shell *shell,
-				  size_t argc, char *argv[])
-{
-	int err;
-	const char *action;
-
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	action = argv[1];
-
-	if (!strcmp(action, "on")) {
-		err = bt_br_set_discoverable(true);
-	} else if (!strcmp(action, "off")) {
-		err = bt_br_set_discoverable(false);
-	} else {
-		shell_help_print(shell, NULL, 0);
-		return;
-	}
-
-	if (err) {
-		printk("BR/EDR set/reset discoverable failed (err %d)\n", err);
-	} else {
-		printk("BR/EDR set/reset discoverable done\n");
-	}
-}
-
-static void cmd_bredr_connectable(const struct shell *shell,
-				 size_t argc, char *argv[])
-{
-	int err;
-	const char *action;
-
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	action = argv[1];
-
-	if (!strcmp(action, "on")) {
-		err = bt_br_set_connectable(true);
-	} else if (!strcmp(action, "off")) {
-		err = bt_br_set_connectable(false);
-	} else {
-		shell_help_print(shell, NULL, 0);
-		return;
-	}
-
-	if (err) {
-		printk("BR/EDR set/rest connectable failed (err %d)\n", err);
-	} else {
-		printk("BR/EDR set/reset connectable done\n");
-	}
-}
-
-static void cmd_bredr_oob(const struct shell *shell,
-			 size_t argc, char *argv[])
-{
-	char addr[BT_ADDR_STR_LEN];
-	struct bt_br_oob oob;
-	int err;
-
-	err = bt_br_oob_get_local(&oob);
-	if (err) {
-		printk("BR/EDR OOB data failed\n");
-		return;
-	}
-
-	bt_addr_to_str(&oob.addr, addr, sizeof(addr));
-
-	printk("BR/EDR OOB data:\n");
-	printk("  addr %s\n", addr);
-}
-
-static void cmd_bredr_sdp_find_record(const struct shell *shell,
-				      size_t argc, char *argv[])
-{
-	int err = 0, res;
-	const char *action;
-
-	if (!default_conn) {
-		printk("Not connected\n");
-		return;
-	}
-
-	if (!shell_cmd_precheck(shell, argc == 2, NULL, 0)) {
-		return;
-	}
-
-	action = argv[1];
-
-	if (!strcmp(action, "HFPAG")) {
-		discov = discov_hfpag;
-	} else if (!strcmp(action, "A2SRC")) {
-		discov = discov_a2src;
-	} else {
-		shell_help_print(shell, NULL, 0);
-		return;
-	}
-
-	if (err) {
-		printk("SDP UUID to resolve not valid (err %d)\n", err);
-		printk("Supported UUID are \'HFPAG\' \'A2SRC\' only\n");
-		return;
-	}
-
-	printk("SDP UUID \'%s\' gets applied\n", action);
-
-	res = bt_sdp_discover(default_conn, &discov);
-	if (res) {
-		printk("SDP discovery failed: result %d\n", res);
-	} else {
-		printk("SDP discovery started\n");
-	}
-}
-#endif
-
 #define HELP_NONE "[none]"
 #define HELP_ADDR_LE "<address: XX:XX:XX:XX:XX:XX> <type: (public|random)>"
 
@@ -2309,9 +1602,6 @@ SHELL_CREATE_STATIC_SUBCMD_SET(bt_cmds) {
 #if defined(CONFIG_BT_FIXED_PASSKEY)
 	SHELL_CMD(fixed-passkey, NULL, "[passkey]", cmd_fixed_passkey),
 #endif
-#if defined(CONFIG_BT_BREDR)
-	SHELL_CMD(auth-pincode, NULL, "<pincode>", cmd_auth_pincode),
-#endif /* CONFIG_BT_BREDR */
 #endif /* CONFIG_BT_SMP || CONFIG_BT_BREDR) */
 #if defined(CONFIG_BT_GATT_CLIENT)
 	SHELL_CMD(gatt-exchange-mtu, NULL, HELP_NONE, cmd_gatt_exchange_mtu),
@@ -2358,24 +1648,6 @@ SHELL_CREATE_STATIC_SUBCMD_SET(bt_cmds) {
 	SHELL_CMD(l2cap-recv, NULL, "[delay (in miliseconds)", cmd_l2cap_recv),
 	SHELL_CMD(l2cap-metrics, NULL, "<value on, off>", cmd_l2cap_metrics),
 #endif
-#if defined(CONFIG_BT_BREDR)
-	SHELL_CMD(br-iscan, NULL, "<value: on, off>", cmd_bredr_discoverable),
-	SHELL_CMD(br-pscan, NULL, "value: on, off", cmd_bredr_connectable),
-	SHELL_CMD(br-connect, NULL, "<address>", cmd_connect_bredr),
-	SHELL_CMD(br-discovery, NULL,
-		  "<value: on, off> [length: 1-48] [mode: limited]",
-		  cmd_bredr_discovery),
-	SHELL_CMD(br-l2cap-register, NULL, "<psm>", cmd_bredr_l2cap_register),
-	SHELL_CMD(br-oob, NULL, NULL, cmd_bredr_oob),
-	SHELL_CMD(br-sdp-find, NULL, "<HFPAG>", cmd_bredr_sdp_find_record),
-#if defined(CONFIG_BT_RFCOMM)
-	SHELL_CMD(br-rfcomm-register, NULL, "<channel>",
-		  cmd_bredr_rfcomm_register),
-	SHELL_CMD(br-rfcomm-connect, NULL, "<channel>", cmd_rfcomm_connect),
-	SHELL_CMD(br-rfcomm-send, NULL, "<number of packets>", cmd_rfcomm_send),
-	SHELL_CMD(br-rfcomm-disconnect, NULL, HELP_NONE, cmd_rfcomm_disconnect),
-#endif /* CONFIG_BT_RFCOMM */
-#endif /* CONFIG_BT_BREDR */
 #endif /* CONFIG_BT_CONN */
 #if defined(CONFIG_BT_CTLR_ADV_EXT)
 	SHELL_CMD(advx, NULL, "<on off> [coded] [anon] [txp]", cmd_advx),
