+++ subsys/bluetooth/shell/bt.h
@@ -22,4 +22,7 @@ extern const struct shell *ctx_shell;
 
 extern struct bt_conn *default_conn;
 
+int str2bt_addr(const char *str, bt_addr_t *addr);
+void conn_addr_str(struct bt_conn *conn, char *addr, size_t len);
+
 #endif /* __BT_H */
